models/requirements.txt
ultralytics
easyocr
opencv-python
torch
torchvision
pillow
numpy
matplotlib
huggingface-hub
onnx
onnxruntime